taskTwoAddedSomeChanges

+++ lab1/task2.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-    int a, b, c, d, e, f, g, h, result;
+    double a, b, c, d, e, f, g, h, result;
     cout << "Введите 8 цифр чтобы получить среднее арифметическое " << endl;
     cin >> a >> b >> c >> d >> e >> f >> g >> h;
     result = (a + b + c + d + e + f + g + h) / 8;
